feat: rv64 1-row precompiles (#483)
* ed add/decompress

* ed-add works

* ed decompress works

* uncomment stuff

* fp works

* fp2 works

* uint256 fails

* uint256 fixed

* u256x2048 works

* weierstrass still need to test

* secp256k1 add decompress

* secp256k1 double

* nit

* assert carry 0

+++ crates/core/machine/src/control_flow/branch/trace.rs
@@ -108,7 +108,7 @@ impl BranchChip {
         let use_signed_comparison = matches!(event.opcode, Opcode::BLT | Opcode::BGE);
 
         let a_lt_b = if use_signed_comparison {
-            (event.a as i32) < (event.b as i32)
+            (event.a as i64) < (event.b as i64)
         } else {
             event.a < event.b
         };
